Initial Navigation Code
To run the initial application on my computer I had to add the line "minSdkVersion 21" in app/build.gradle and change the version of kotlin in the settings.gradle to the latest version. Hopefully, if you have to change this back for it to run on your computer it won't hinder the navigation code. Currently, the navigation itself is basic: you have a button that constantly updates your real location with an origin marker and holding down on the screen will create a destination marker. As long as a possible path is between the two markers a polyline will be made along with a given distance and estimated time. Constantly updating the origin marker will change the distance, time, and polyline length accordingly. This code includes a directions model that keeps track and creates step-by-step instructions whenever a path is made. I plan on using the html instruction data, along with the web-socket server, to display these directions onto the glasses, just as we can with the transcription, giving our glasses very basic navigation capabilities.